pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "composing-functions"
version = "1.0.0"
description = "Codebase for the paper 'How Do Language Models Compose Functions?'"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{name = "Apoorv Khandelwal", email = "mail@apoorvkh.com"}]
dependencies = [
    "torch==2.6.0",
    "einops==0.8.1",
    "transformers[hf-xet]==4.52.4",
    "nnsight==0.4.6",
    # data utils
    "easynmt==2.0.2",
    "fasttext-wheel==0.9.2",
    "sacremoses==0.1.1",
    "num2words==0.5.14",
    "webcolors==24.11.1",
    # visualization
    "marimo[lsp]==0.15.3",
    "polars[style]==1.22.0",
    "altair[save]==5.5.0",
    # interface
    "tyro==0.9.31",
    "ai2-tango==1.3.2",
    # linting, formatting, type checking
    "ruff==0.13.0",
    "pyright[nodejs]==1.1.405",
]

[tool.uv]
override-dependencies = [
    "fasttext ; sys_platform == 'never'",  # using pre-built wheels via fasttext-wheel instead
]

[tool.ruff]
line-length = 120
src = ["src", "plotting", "configs"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pyright]
include = ["src", "plotting", "configs"]
pythonVersion = "3.12"
